rename file and change to use builtin method for wrappring
