+++ qt/python/mantidqt/widgets/instrumentview/api.py
@@ -1,0 +1,61 @@
+#!/usr/bin/env python
+# Mantid Repository : https://github.com/mantidproject/mantid
+#
+# Copyright &copy; 2017 ISIS Rutherford Appleton Laboratory UKRI,
+#   NScD Oak Ridge National Laboratory, European Spallation Source,
+#   Institut Laue - Langevin & CSNS, Institute of High Energy Physics, CAS
+# SPDX - License - Identifier: GPL - 3.0 +
+#  This file is part of the mantidqt package
+#
+
+import functools
+from mantidqt.utils.qt.qappthreadcall import QAppThreadCall
+from mantidqt.widgets.instrumentview.presenter import InstrumentViewPresenter
+from mantidqt.utils.qt.qappthreadcall import force_method_calls_to_qapp_thread
+
+
+def safe_qthread(func):
+    """decorator function to move given call to main QappThread"""
+    @functools.wraps(func)
+    def _wrapped(*args, **kwargs):
+        return QAppThreadCall(func)(*args, **kwargs)
+
+    return _wrapped
+
+
+def get_instrumentview(workspace):
+    """Return a handle to the instrument view of given workspace
+    :param ws: input workspace
+    """
+    def _wrappper(ws):
+        return force_method_calls_to_qapp_thread(InstrumentViewPresenter(ws))
+
+    # need to do some duck-typing here
+    ivp = QAppThreadCall(_wrappper)(workspace)
+    # link nested method to top level
+    # NOTE: setMin and setMax still leads to segfault, need to force
+    #       wrapped in QAppThreadCall again
+    ivp.select_tab = ivp.container.select_tab
+    ivp.select_surface_type = ivp.get_render_tab().setSurfaceType
+    ivp.set_min_value = safe_qthread(ivp.get_render_tab().setMinValue)
+    ivp.set_max_value = safe_qthread(ivp.get_render_tab().setMaxValue)
+    ivp.set_auto_scaling = ivp.get_render_tab().setColorMapAutoscaling
+    ivp.set_axis = ivp.get_render_tab().setAxis
+    return ivp
+
+
+if __name__ == "__main__":
+    import sys
+    from mantidqt.gui_helper import get_qapplication
+    from mantid.simpleapi import LoadEventNexus
+
+    app, within_mantid = get_qapplication()
+
+    nexus_path = ("/SNS/EQSANS/shared/sans-backend/data/new/ornl/sans/hfir/gpsans/CG2_9177.nxs.h5")
+    ws = LoadEventNexus(Filename=nexus_path, NumberOfBins=10)
+
+    myiv = get_instrumentview(ws)
+    myiv.show_view()
+
+    if not within_mantid:
+        sys.exit(app.exec_())
